% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketchedRF-package.R
\docType{package}
\name{sketchedRF-package}
\alias{sketchedRF-package}
\title{Fit Sketched Random Forest}
\description{
Classification based on a forest of trees trained on multiple sets of data perturbed via matrix sketching.
}
\seealso{
\link{sketchedRF}, \link{predict.sketchedRF}
}
\author{
Laura Anderlucci \email{laura.anderlucci@unibo.it}

Angela Montanari \email{angela.montanari@unibo.it}
}
\keyword{package}
