% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketchedRF.R
\name{sketchedRF}
\alias{sketchedRF}
\title{Sketched Random Forest}
\usage{
sketchedRF(
  x = x,
  y = y,
  nk = NULL,
  sketch = "Gaussian",
  ntree = 500,
  OOB = TRUE,
  nk.OOB = NULL,
  seed = 1
)
}
\arguments{
\item{x}{Data matrix of predictors.}

\item{y}{Vector of responses.}

\item{nk}{Vector of desired classes' size after sketching. By default, it is set as equal for all the classes, and equal to size of the largest class in the training set.}

\item{sketch}{Type of sketching matrix, one among \code{"Gaussian"},\code{"CW"}, \code{"Hadamard"}.}

\item{ntree}{no. of trees in the forest. Default is 500.}

\item{OOB}{Return sketched out-of-bag error estimate. Default is TRUE.}

\item{nk.OOB}{Desired classes' size for the out-of-bag sample. By default it is equal to the training set class distribution.}

\item{seed}{Set the seed to obtain reproducible results.}
}
\value{
An object of class \code{sketchedRF}, which is a list with the following components:
\item{call}{the original call to randomForest}
\item{ntree}{number of trees grown}
\item{mtry}{number of predictors sampled for spliting at each node.}
\item{sketch}{type ok employed sketching matrices}
\item{sOOB.err}{the sketched out-of-bag estimate of  error (when \code{OOB=TRUE}).}
}
\description{
Sketched Random Forest, proposed by Anderlucci, Montanari (2025), yields a forest of classification trees trained on multiple sets of data perturbed via matrix sketching.
}
\examples{
train<-c(1:45,51:95,101:145)
test<-c(46:50,96:100,146:150)
out.sRF<-sketchedRF(x=iris[train,-5],y=iris[train,5],nk=NULL,sketch="Gaussian",ntree=500,OOB=TRUE)
}
\references{
L. Anderlucci, A. Montanari, M. Ferracin (2025). \emph{Perturbing data to address dataset shift in metastatic cancer classification}.
}
\seealso{
\link{predict.sketchedRF}, \link{varImp}, \link{varImpPlot.sketchedRF}
}
