% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp.R
\name{varImp}
\alias{varImp}
\title{Extract variable importance measure}
\usage{
varImp(object, type = "all", nk = NULL, seed = 1)
}
\arguments{
\item{object}{an object of class skecthedRF, as that created by the function sketchedRF.}

\item{type}{one of 1, 2, "all", indicating the variable importance measure: \code{1}=mean decrease in accuracy, \code{2}=mean decrease in node impurity, \code{"all"} both of them.}

\item{nk}{Vector of desired classes' sizes for the sketched OOB sample. By default is the same of the training set.}

\item{seed}{Set the random seed for the generation of sketched OOB data.}
}
\value{
A matrix of importance measure, one row for each predictor variable. The column(s) are different importance measures.
}
\description{
This is the function to compute variable importance measures as produced by sketched random Forest.
}
\details{
Here are the definitions of the variable importance measures. The first measure is computed from permuting OOB data: For each tree, the prediction error on the out-of-bag portion of the data is recorded (error rate for classification). Then the same is done after permuting each predictor variable. The difference between the two are then averaged over all trees, and normalized by the standard deviation of the differences. The second measure is the total decrease in node impurities from splitting on the variable, averaged over all trees. The node impurity is measured by the Gini index.
}
\examples{
train<-c(1:45,51:95,101:145)
test<-c(46:50,96:100,146:150)
out.sRF<-sketchedRF(x=iris[train,-5],y=iris[train,5],nk=NULL,sketch="Gaussian",ntree=500)
varImp(out.sRF,type="all")
}
\references{
L. Anderlucci, A. Montanari, M. Ferracin (2025). \emph{Perturbing data to address dataset shift in metastatic cancer classification}.
}
\seealso{
\link{sketchedRF}, \link{predict.sketchedRF}, \link{varImpPlot.sketchedRF}
}
