% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImpPlot.sketchedRF.R
\name{varImpPlot.sketchedRF}
\alias{varImpPlot.sketchedRF}
\title{Variable Importance Plot}
\usage{
varImpPlot.sketchedRF(
  imp_obj,
  sort = TRUE,
  n.var = min(30, nrow(imp_obj)),
  main = deparse(substitute(imp_obj)),
  ...
)
}
\arguments{
\item{imp_obj}{an object of class sketchedRF, as that created by the function \code{importance}.}

\item{sort}{Should the variables be sorted in decreasing order of importance?}

\item{n.var}{How many variables to show? (Ignored if sort=FALSE.)}

\item{main}{-plot title.}

\item{...}{Other graphical parameters to be passed on to dotchart.}
}
\value{
Invisibly, the importance of the variables that were plotted.
}
\description{
Dotchart of variable importance as measured by a sketched Random Forest.
}
\examples{
train<-c(1:45,51:95,101:145)
test<-c(46:50,96:100,146:150)
out.sRF<-sketchedRF(x=iris[train,-5],y=iris[train,5],nk=NULL,sketch="Gaussian",ntree=500)
iris_import<-varImp(out.sRF,type="all")
varImpPlot.sketchedRF(iris_import)
}
\seealso{
\link{sketchedRF}, \link{predict.sketchedRF}, \link{varImp}
}
